+++ pilot/pkg/networking/util/util_test.go
@@ -20,14 +20,12 @@ import (
 	"time"
 
 	v2 "github.com/envoyproxy/go-control-plane/envoy/api/v2"
-	envoyauth "github.com/envoyproxy/go-control-plane/envoy/api/v2/auth"
 	core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
 	endpoint "github.com/envoyproxy/go-control-plane/envoy/api/v2/endpoint"
 	listener "github.com/envoyproxy/go-control-plane/envoy/api/v2/listener"
 	http_conn "github.com/envoyproxy/go-control-plane/envoy/config/filter/network/http_connection_manager/v2"
 	xdsutil "github.com/envoyproxy/go-control-plane/pkg/wellknown"
 
-	"github.com/davecgh/go-spew/spew"
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes"
 	"github.com/golang/protobuf/ptypes/any"
@@ -37,9 +35,7 @@ import (
 
 	networking "istio.io/api/networking/v1alpha3"
 
-	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pilot/pkg/model"
-	authn_model "istio.io/istio/pilot/pkg/security/model"
 	"istio.io/istio/pilot/pkg/serviceregistry"
 	proto2 "istio.io/istio/pkg/proto"
 )
@@ -808,149 +804,6 @@ func TestBuildStatPrefix(t *testing.T) {
 			got := BuildStatPrefix(tt.statPattern, tt.host, tt.subsetName, tt.port, tt.attributes)
 			if got != tt.want {
 				t.Errorf("Expected alt statname %s, but got %s", tt.want, got)
-			}
-		})
-	}
-}
-
-func TestApplyToCommonTLSContext(t *testing.T) {
-	testCases := []struct {
-		name       string
-		sdsUdsPath string
-		node       *model.Proxy
-		result     *envoyauth.CommonTlsContext
-	}{
-		{
-			name:       "MTLSStrict using SDS",
-			sdsUdsPath: "/tmp/sdsuds.sock",
-			node: &model.Proxy{
-				Metadata: &model.NodeMetadata{
-					SdsEnabled: true,
-				},
-			},
-			result: &envoyauth.CommonTlsContext{
-				TlsCertificateSdsSecretConfigs: []*envoyauth.SdsSecretConfig{
-					{
-						Name: "default",
-						SdsConfig: &core.ConfigSource{
-							InitialFetchTimeout: features.InitialFetchTimeout,
-							ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-								ApiConfigSource: &core.ApiConfigSource{
-									ApiType: core.ApiConfigSource_GRPC,
-									GrpcServices: []*core.GrpcService{
-										{
-											TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-												EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: authn_model.SDSClusterName},
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-				ValidationContextType: &envoyauth.CommonTlsContext_CombinedValidationContext{
-					CombinedValidationContext: &envoyauth.CommonTlsContext_CombinedCertificateValidationContext{
-						DefaultValidationContext: &envoyauth.CertificateValidationContext{MatchSubjectAltNames: StringToExactMatch([]string{})},
-						ValidationContextSdsSecretConfig: &envoyauth.SdsSecretConfig{
-							Name: "ROOTCA",
-							SdsConfig: &core.ConfigSource{
-								InitialFetchTimeout: features.InitialFetchTimeout,
-								ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-									ApiConfigSource: &core.ApiConfigSource{
-										ApiType: core.ApiConfigSource_GRPC,
-										GrpcServices: []*core.GrpcService{
-											{
-												TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-													EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: authn_model.SDSClusterName},
-												},
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name:       "ISTIO_MUTUAL SDS without node meta",
-			sdsUdsPath: "/tmp/sdsuds.sock",
-			node: &model.Proxy{
-				Metadata: &model.NodeMetadata{},
-			},
-			result: &envoyauth.CommonTlsContext{
-				TlsCertificates: []*envoyauth.TlsCertificate{
-					{
-						CertificateChain: &core.DataSource{
-							Specifier: &core.DataSource_Filename{
-								Filename: "/etc/certs/cert-chain.pem",
-							},
-						},
-						PrivateKey: &core.DataSource{
-							Specifier: &core.DataSource_Filename{
-								Filename: "/etc/certs/key.pem",
-							},
-						},
-					},
-				},
-				ValidationContextType: &envoyauth.CommonTlsContext_ValidationContext{
-					ValidationContext: &envoyauth.CertificateValidationContext{
-						TrustedCa: &core.DataSource{
-							Specifier: &core.DataSource_Filename{
-								Filename: "/etc/certs/root-cert.pem",
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name:       "ISTIO_MUTUAL with custom cert paths from proxy node metadata",
-			sdsUdsPath: "/tmp/sdsuds.sock",
-			node: &model.Proxy{
-				Metadata: &model.NodeMetadata{
-					TLSServerCertChain: "/custom/path/to/cert-chain.pem",
-					TLSServerKey:       "/custom-key.pem",
-					TLSServerRootCert:  "/custom/path/to/root.pem",
-				},
-			},
-			result: &envoyauth.CommonTlsContext{
-				TlsCertificates: []*envoyauth.TlsCertificate{
-					{
-						CertificateChain: &core.DataSource{
-							Specifier: &core.DataSource_Filename{
-								Filename: "/custom/path/to/cert-chain.pem",
-							},
-						},
-						PrivateKey: &core.DataSource{
-							Specifier: &core.DataSource_Filename{
-								Filename: "/custom-key.pem",
-							},
-						},
-					},
-				},
-				ValidationContextType: &envoyauth.CommonTlsContext_ValidationContext{
-					ValidationContext: &envoyauth.CertificateValidationContext{
-						TrustedCa: &core.DataSource{
-							Specifier: &core.DataSource_Filename{
-								Filename: "/custom/path/to/root.pem",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, test := range testCases {
-		t.Run(test.name, func(t *testing.T) {
-			tlsContext := &envoyauth.CommonTlsContext{}
-			ApplyToCommonTLSContext(tlsContext, test.node.Metadata, test.sdsUdsPath, []string{})
-
-			if !reflect.DeepEqual(tlsContext, test.result) {
-				t.Errorf("got() = %v, want %v", spew.Sdump(tlsContext), spew.Sdump(test.result))
 			}
 		})
 	}
